+++ environments/docker/install.sh
@@ -12,13 +12,9 @@ else
   NOEXTRA=""
 fi
 
-# install uv and add the executable to the PATH
-wget -qO- https://astral.sh/uv/0.7.14/install.sh | sh
-source $HOME/.local/bin/env
-
 # install all requirements into the existing system environment
 uv export --no-cache --no-build --all-extras $NOEXTRA --group examples > requirements-temp.txt
-uv pip install --no-cache --no-build --system --strict -r requirements-temp.txt
+uv pip install --no-cache --no-build --system --compile-bytecode --strict -r requirements-temp.txt
 rm -f requirements-temp.txt
 uv cache clean
 python --version